tests/wpt/web-platform-tests/html/semantics/document-metadata/the-style-element/style_non_matching_media.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: Non-matching media type should have stylesheet</title>
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-style-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style media="unknown">
      body { color: green }
    </style>
  </head>
  <body>
    <script>
      test(function() {
        assert_equals(document.styleSheets.length, 1);
      });
    </script>
  </body>
</html>
