Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-style-element/style_non_matching_media-expected.txt


FAIL HTML Test: Non-matching media type should have stylesheet assert_equals: expected 1 but got 0
